==> STEP3-NLP.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
from nltk.corpus import brown
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer
from nltk.sentiment import SentimentIntensityAnalyzer
from nltk.corpus import wordnet
from collections import  Counter
import spacy
from spacy import displacy
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

df=pd.read_csv('DATA_After_Cleaning.csv')

count_vec = CountVectorizer(stop_words='english',ngram_range=(1, 2), max_features= 7)
count_vec.fit(df['Updated_Review'])
text_vectorized = pd.DataFrame(count_vec.transform(df['Updated_Review']).toarray(),
                               columns=count_vec.get_feature_names())
text_vectorized

final_df = df.merge(text_vectorized,left_index=True, right_index=True)

final_df.to_csv('After_cleaning.csv',index=False)
